January 11, 2022 3:29 PM

+++ utils_old.py
@@ -149,7 +149,7 @@ def save_tensor_as_img(tensor, save_path):
     print(tensor.shape)
     tensor = tensor.permute(0, 3, 1, 2)
     print(tensor.shape)
-    TF.to_pil_image(tensor[0].cpu()).save(save_path)
+    TF.to_pil_image(tensor[0].cpu()).astype(np.uint8).save(save_path)
 
 
 class MakeCutoutsDet(nn.Module):
